requirements.txt
asgiref==3.8.1
Django==5.1.6
django-simple-history==3.8.0
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
pillow==11.1.0
python-dotenv==1.0.1
dj-database-url-2.3.0
python-dotenv==1.0.0
django-filter==23.5

